+++ fs/nxffs/nxffs_read.c
@@ -1,37 +1,20 @@
 /****************************************************************************
  * fs/nxffs/nxffs_read.c
  *
- *   Copyright (C) 2011, 2013, 2017-2018 Gregory Nutt. All rights reserved.
- *   Author: Gregory Nutt <gnutt@nuttx.org>
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.  The
+ * ASF licenses this file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the
+ * License.  You may obtain a copy of the License at
  *
- * References: Linux/Documentation/filesystems/romfs.txt
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in
- *    the documentation and/or other materials provided with the
- *    distribution.
- * 3. Neither the name NuttX nor the names of its contributors may be
- *    used to endorse or promote products derived from this software
- *    without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
- * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
- * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
- * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
- * AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
- * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
  *
  ****************************************************************************/
 
@@ -41,6 +24,7 @@
 
 #include <nuttx/config.h>
 
+#include <stdint.h>
 #include <string.h>
 #include <fcntl.h>
 #include <crc32.h>
@@ -62,7 +46,7 @@
  *
  * Description:
  *   Seek to the file position before read or write access.  Note that the
- *   simplier nxffs_ioseek() cannot be used for this purpose.  File offsets
+ *   simpler nxffs_ioseek() cannot be used for this purpose.  File offsets
  *   are not easily mapped to FLASH offsets due to intervening block and
  *   data headers.
  *
@@ -101,7 +85,9 @@ static ssize_t nxffs_rdseek(FAR struct nxffs_volume_s *volume,
   datend = 0;
   do
     {
-      /* Check if the next data block contains the sought after file position */
+      /* Check if the next data block contains the sought after file
+       * position
+       */
 
       ret = nxffs_nextblock(volume, offset, blkentry);
       if (ret < 0)
@@ -124,13 +110,15 @@ static ssize_t nxffs_rdseek(FAR struct nxffs_volume_s *volume,
   /* Return the offset to the data within the current data block */
 
   blkentry->foffset = fpos - datstart;
-  nxffs_ioseek(volume, blkentry->hoffset + SIZEOF_NXFFS_DATA_HDR + blkentry->foffset);
+  nxffs_ioseek(volume, blkentry->hoffset + SIZEOF_NXFFS_DATA_HDR +
+               blkentry->foffset);
   return OK;
 }
 
 /****************************************************************************
  * Public Functions
  ****************************************************************************/
+
 /****************************************************************************
  * Name: nxffs_read
  *
@@ -150,7 +138,7 @@ ssize_t nxffs_read(FAR struct file *filep, FAR char *buffer, size_t buflen)
   size_t readsize;
   int ret;
 
-  finfo("Read %d bytes from offset %d\n", buflen, filep->f_pos);
+  finfo("Read %zu bytes from offset %jd\n", buflen, (intmax_t)filep->f_pos);
 
   /* Sanity checks */
 
@@ -321,7 +309,8 @@ int nxffs_nextblock(FAR struct nxffs_volume_s *volume, off_t offset,
           if (ch != g_datamagic[nmagic])
             {
               /* Ooops... this is the not the right character for the magic
-               * Sequence.  Check if we need to restart or to cancel the sequence:
+               * Sequence.  Check if we need to restart or to cancel the
+               * sequence:
                */
 
               if (ch == g_datamagic[0])
@@ -354,11 +343,12 @@ int nxffs_nextblock(FAR struct nxffs_volume_s *volume, off_t offset,
 
               /* Read the block header and verify the block at that address */
 
-              ret = nxffs_rdblkhdr(volume, blkentry->hoffset, &blkentry->datlen);
+              ret = nxffs_rdblkhdr(volume, blkentry->hoffset,
+                                   &blkentry->datlen);
               if (ret == OK)
                 {
-                  finfo("Found a valid data block, offset: %d datlen: %d\n",
-                        blkentry->hoffset, blkentry->datlen);
+                  finfo("Found a valid data block, offset: %jd datlen: %d\n",
+                        (intmax_t)blkentry->hoffset, blkentry->datlen);
                   return OK;
                 }
 
@@ -385,8 +375,8 @@ int nxffs_nextblock(FAR struct nxffs_volume_s *volume, off_t offset,
  *
  * Input Parameters:
  *   volume - Describes the current volume.
- *   offset - The byte offset from the beginning of FLASH where the data block
- *     header is expected.
+ *   offset - The byte offset from the beginning of FLASH where the data
+ *     block header is expected.
  *   datlen  - A memory location to return the data block length.
  *
  * Returned Value:
@@ -405,7 +395,9 @@ int nxffs_rdblkhdr(FAR struct nxffs_volume_s *volume, off_t offset,
   uint16_t dlen;
   int ret;
 
-  /* Make sure that the block containing the data block header is in the cache */
+  /* Make sure that the block containing the data block header is in the
+   * cache
+   */
 
   nxffs_ioseek(volume, offset);
   ret = nxffs_rdcache(volume, volume->ioblock);
@@ -432,7 +424,8 @@ int nxffs_rdblkhdr(FAR struct nxffs_volume_s *volume, off_t offset,
 
   if ((uint32_t)doffset + (uint32_t)dlen > (uint32_t)volume->geo.blocksize)
     {
-      ferr("ERROR: Data length=%d is unreasonable at offset=%d\n", dlen, doffset);
+      ferr("ERROR: Data length=%d is unreasonable at offset=%d\n", dlen,
+           doffset);
       return -EIO;
     }
