+++ fs/nxffs/nxffs_write.c
@@ -1,37 +1,20 @@
 /****************************************************************************
  * fs/nxffs/nxffs_write.c
  *
- *   Copyright (C) 2011, 2013, 2017-2018 Gregory Nutt. All rights reserved.
- *   Author: Gregory Nutt <gnutt@nuttx.org>
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.  The
+ * ASF licenses this file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the
+ * License.  You may obtain a copy of the License at
  *
- * References: Linux/Documentation/filesystems/romfs.txt
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in
- *    the documentation and/or other materials provided with the
- *    distribution.
- * 3. Neither the name NuttX nor the names of its contributors may be
- *    used to endorse or promote products derived from this software
- *    without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
- * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
- * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
- * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
- * AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
- * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
  *
  ****************************************************************************/
 
@@ -41,6 +24,7 @@
 
 #include <nuttx/config.h>
 
+#include <stdint.h>
 #include <string.h>
 #include <fcntl.h>
 #include <crc32.h>
@@ -227,7 +211,9 @@ static inline int nxffs_wralloc(FAR struct nxffs_volume_s *volume,
           ret = nxffs_hdrerased(volume, wrfile, mindata);
           if (ret == OK)
             {
-              /* Valid memory for the data block was found.  Return success. */
+              /* Valid memory for the data block was found.  Return
+               * success.
+               */
 
               return OK;
             }
@@ -519,7 +505,8 @@ static inline ssize_t nxffs_zappend(FAR struct nxffs_volume_s *volume,
  *
  ****************************************************************************/
 
-ssize_t nxffs_write(FAR struct file *filep, FAR const char *buffer, size_t buflen)
+ssize_t nxffs_write(FAR struct file *filep, FAR const char *buffer,
+                    size_t buflen)
 {
   FAR struct nxffs_volume_s *volume;
   FAR struct nxffs_wrfile_s *wrfile;
@@ -528,7 +515,7 @@ ssize_t nxffs_write(FAR struct file *filep, FAR const char *buffer, size_t bufle
   ssize_t total;
   int ret;
 
-  finfo("Write %d bytes to offset %d\n", buflen, filep->f_pos);
+  finfo("Write %zd bytes to offset %jd\n", buflen, (intmax_t)filep->f_pos);
 
   /* Sanity checks */
 
@@ -590,7 +577,9 @@ ssize_t nxffs_write(FAR struct file *filep, FAR const char *buffer, size_t bufle
 
       nxffs_ioseek(volume, wrfile->doffset);
 
-      /* Verify that the FLASH data that was previously written is still intact */
+      /* Verify that the FLASH data that was previously written is still
+       * intact
+       */
 
       ret = nxffs_reverify(volume, wrfile);
       if (ret < 0)
@@ -606,7 +595,8 @@ ssize_t nxffs_write(FAR struct file *filep, FAR const char *buffer, size_t bufle
       nwritten = nxffs_wrappend(volume, wrfile, &buffer[total], remaining);
       if (nwritten < 0)
         {
-          ferr("ERROR: Failed to append to FLASH to a data block: %d\n", -ret);
+          ferr("ERROR: Failed to append to FLASH to a data block: %d\n",
+               -ret);
           goto errout_with_semaphore;
         }
 
@@ -684,7 +674,9 @@ int nxffs_wrextend(FAR struct nxffs_volume_s *volume,
 
       nxffs_ioseek(volume, wrfile->doffset);
 
-      /* Verify that the FLASH data that was previously written is still intact */
+      /* Verify that the FLASH data that was previously written is still
+       * intact
+       */
 
       ret = nxffs_reverify(volume, wrfile);
       if (ret < 0)
@@ -741,7 +733,7 @@ int nxffs_wrextend(FAR struct nxffs_volume_s *volume,
  *   On successful return the following are also valid:
  *
  *   volume->ioblock - Read/write block number of the block containing the
- *     candidate oject position
+ *     candidate object position
  *   volume->iooffset - The offset in the block to the candidate object
  *     position.
  *   volume->froffset - Updated offset to the first free FLASH block after
@@ -828,13 +820,13 @@ int nxffs_wrreserve(FAR struct nxffs_volume_s *volume, size_t size)
  *   following settings (left by nxffs_wrreserve()):
  *
  *   volume->ioblock - Read/write block number of the block containing the
- *     candidate oject position
+ *     candidate object position
  *   volume->iooffset - The offset in the block to the candidate object
  *     position.
  *
  * Input Parameters:
  *   volume - Describes the NXFFS volume
- *   size - The size of the object to be verifed.
+ *   size - The size of the object to be verified.
  *
  * Returned Value:
  *   Zero is returned on success.  Otherwise, a negated errno value is
@@ -873,8 +865,8 @@ int nxffs_wrverify(FAR struct nxffs_volume_s *volume, size_t size)
            * the block has uncorrectable bit errors.
            */
 
-          ferr("ERROR: Failed to read block %d: %d\n",
-               volume->ioblock, -ret);
+          ferr("ERROR: Failed to read block %jd: %d\n",
+               (intmax_t)volume->ioblock, -ret);
         }
 
       /* Search to the very end of this block if we have to */
@@ -898,7 +890,7 @@ int nxffs_wrverify(FAR struct nxffs_volume_s *volume, size_t size)
 
                   if (nerased >= size)
                     {
-                      /* Yes.. this this is where we will put the object */
+                      /* Yes.. this is where we will put the object */
 
                       off_t offset =
                         volume->ioblock * volume->geo.blocksize + iooffset;
@@ -1005,7 +997,8 @@ int nxffs_wrblkhdr(FAR struct nxffs_volume_s *volume,
    *     begin the search for the next inode header or data block.
    */
 
-  volume->froffset = (wrfile->doffset + wrfile->datlen + SIZEOF_NXFFS_DATA_HDR);
+  volume->froffset = (wrfile->doffset + wrfile->datlen +
+                      SIZEOF_NXFFS_DATA_HDR);
 
   /* wrfile->file.entry:
    *   datlen:  Total file length accumulated so far.  When the file is
